# restapi/static/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MANTIS API",
    "version": "1.0.1",
    "description": "MANTIS API.\n\n## Authentication\n\nProtected endpoints require a valid **Bearer token**.\n\n### Step 1 \u2014 Obtain an access token\n\nSend a **POST** request to **`/authentication/token`** with:\n\n```json\n{\n  \"userId\": \"user@sts.my\",\n  \"accessKey\": \"<your-access-key>\"\n}\n```\n\n**Example response (200 OK):**\n\n```json\n{\n  \"accessToken\": \"<jwt-token>\",\n  \"expiredDate\": \"2026-01-01 12:00:00\"\n}\n```\n\n### Step 2 \u2014 Authorize in Swagger UI\n\n1. Call **POST /authentication/token** and copy the `accessToken`.\n2. Click **Authorize** (lock icon).\n3. Paste the `accessToken` into **BearerAuth** and click **Authorize**.\n\n### Using the token with external clients\n\n```\nAuthorization: Bearer <accessToken>\n```\n\n## Swagger sample cap (all list endpoints)\n\n**Try it out from this page** is detected via the `Referer` header (`/swagger`). Each list in the JSON is capped at **20 sample rows** so the UI does not lag. The body includes `sample: true`, `fromSwagger: true`, `totalCount` (or `totalCounts`), and a `message`. curl, the operations frontend, and MCP receive the **full** result. `POST /authentication/token` and `GET /` (health) are not sampled.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from **POST /authentication/token**. Enter the `accessToken` value only (Swagger adds the `Bearer` prefix)."
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "userId",
          "accessKey"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "description": "The ID of the user.",
            "example": "user@sts.my"
          },
          "accessKey": {
            "type": "string",
            "description": "The access key given to the user."
          }
        }
      },
      "UserToken": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "The JWT access token."
          },
          "expiredDate": {
            "type": "string",
            "description": "Token expiry datetime (yyyy-MM-dd HH:mm:ss).",
            "example": "2026-12-31 23:59:59"
          }
        }
      },
...
      "DarkVesselsResponse": {
        "type": "object",
        "properties": {
          "ruleVersion": {
            "type": "string",
            "example": "v1.2-slowmove-dark-polygon-ofac-label"
          },
          "shipTypeFilter": {
            "type": "string",
            "example": "70-89 (cargo/tanker/container Class-A large vessels)"
          },
          "minSilenceMinutes": {
            "type": "integer",
            "example": 30
          },
          "coverageExitDays": {
            "type": "integer",
            "example": 3
          },
          "includeCoverageExit": {
            "type": "boolean"
          },
          "candidateCount": {
            "type": "integer"
          },
          "byReason": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "byConfidence": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "sanctionsMatchCount": {
            "type": "integer",
            "description": "Candidates that match OFAC. Unmatched vessels are still returned."
          },
          "vessels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DarkVessel"
            }
          }
        }
      },
...
      "SanctionsListResponse": {
        "type": "object",
        "properties": {
          "imo": {
            "type": "string",
            "nullable": true,
            "description": "Normalized IMO used for the search, if provided."
          },
          "mmsi": {
            "type": "string",
            "nullable": true,
            "description": "Normalized MMSI used for the search, if provided."
          },
          "totalCount": {
            "type": "integer",
            "description": "Rows matching the filter (before Swagger sample cap)."
          },
          "returnedCount": {
            "type": "integer",
            "description": "Rows in `vessels`. 20 when this call is from Swagger UI and the list is larger."
          },
          "sample": {
            "type": "boolean",
            "description": "True when the payload was capped for Swagger UI."
          },
          "sampleLimit": {
            "type": "integer",
            "nullable": true,
            "example": 20
          },
          "fromSwagger": {
            "type": "boolean",
            "description": "True when the request Referer is the Swagger page."
          },
          "message": {
            "type": "string",
            "nullable": true,
            "description": "Explains the 20-row sample when `sample` is true."
          },
          "vessels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfacVessel"
            }
          }
        }
      },
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/authentication/token": {
      "post": {
        "tags": [
          "Token"
        ],
        "summary": "Get the authentication access token.",
        "operationId": "getAccessToken",
        "requestBody": {
          "description": "Credentials required to obtain a token.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserToken"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized \u2014 invalid credentials."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/mantis/darkvessels": {
      "get": {
        "tags": [
          "mantis"
        ],
        "summary": "Get suspected dark / AIS-transponder-off vessels.",
        "description": "v1 heuristic over `ais_vesselslowmoveactivities`: Class-A large vessels (shipType 70\u201389) that slowed then went silent before a confirmed stop (`rowcount < 30`), with AIS silence \u2265 30 minutes. Reasons: `suspected_dark_after_slowdown` (higher interest), `possible_coverage_exit` (SEA footprint leave), `low_evidence_ais_gap`. Not proof of intentional dark \u2014 see `backend/mantis-detection.md`. Candidates are labeled with `polygonName` when the last position is inside a polygon from `polygons.py` (Excl name preferred if inside a hole). They are not dropped by polygon. OFAC SDN/CONS is attached as `sanctionsMatch` / `matchConfidence` (IMO = confirmed, MMSI-only = possible). Unmatched vessels stay. Listed ships are sorted first; dark `confidence` is unchanged.\n\n**This Swagger page:** `vessels` is capped at 20 samples.",
        "operationId": "getDarkVessels",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "includeCoverageExit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "When false, exclude `possible_coverage_exit` rows (tighter ops list)."
          }
        ],
        "responses": {
          "200": {
            "description": "Suspected dark-vessel candidates.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DarkVesselsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized \u2014 missing or invalid Bearer token."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mantis/sanctions": {
      "get": {
        "tags": [
          "mantis"
        ],
        "summary": "Search the OFAC vessel list (IMO or MMSI).",
        "description": "Returns OFAC **vessels** from `pnav` (SDN plus consolidated non-SDN if loaded). This is a lookup of the ingested list, not a MANTIS detector and not a legal finding.\n\n- Pass `imo` and/or `mmsi` to search (either identifier matches).\n- If neither is passed, the **full vessel list** is returned (~1,500 SDN ships).\n- **Swagger UI Try it out** is capped at **20 sample rows** so the page does not lag. `sample` is true and `message` explains this. `totalCount` is still the full match count.\n- curl, the frontend, and MCP receive the entire result.\n\nPeople and companies on SDN are not included (no IMO/MMSI).",
        "operationId": "getSanctionsList",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "imo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "9187631"
            },
            "description": "IMO number (7 digits, or `IMO 9187631`). Optional."
          },
          {
            "name": "mmsi",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "256845000"
            },
            "description": "MMSI. Optional. If both imo and mmsi are set, a row matching either is returned."
          }
        ],
        "responses": {
          "200": {
            "description": "OFAC vessel rows. From Swagger UI, at most 20 rows plus a sample message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SanctionsListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized \u2014 missing or invalid Bearer token."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mantis/polygons": {
      "get": {
        "tags": [
          "mantis"
        ],
        "summary": "Get all anchorage polygons (including restricted limit).",
        "description": "Returns every anchorage area polygon defined in `polygons.py`, plus the **Restricted Limit** polygon. Polygons whose name contains **Excl** are carve-outs (holes) inside a parent polygon and are included in this catalogue. Coordinates are GeoJSON-ordered `[longitude, latitude]`.\n\n**Clients (curl / frontend):** a JSON array of all polygons.\n**This Swagger page:** an object with `polygons` (first 20), `sample`, `totalCount`, and `message`.",
        "operationId": "getAllAnchoragePolygons",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Full polygon catalogue (array) for curl/frontend. From this Swagger page: object with 20 sample polygons.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnchoragePolygon"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized \u2014 missing or invalid Bearer token."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mantis/sts-activities": {
      "get": {
        "tags": [
          "mantis"
        ],
        "summary": "Get active high-suspicion STS pairs inside anchorage polygons.",
        "description": "Returns vessel pairs from **active** (`is_open = true`) proximity clusters with **high suspicion_score** (default \u2265 4.5) whose centroid falls inside a parent / watch polygon. Clusters whose centroid is inside an **Excl** carve-out (hole) are excluded. Pairs are recomputed at \u2264 **35 m**. Each vessel includes **`lastSeenAt`** (AIS time of the fix used for that member). **`pairedAt`** is pipeline wall clock only — use `lastSeenAt` with GET /mantis/vessel-track. OFAC labels (`sanctionsMatch`) are attached per vessel; `suspicionScore` is unchanged. Listed pairs are sorted first.\n\n**This Swagger page:** `pairs` and `pairedVessels` are each capped at 20 samples.",
        "operationId": "getAnchorageStsActivities",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "minSuspicionScore",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double",
              "default": 4.5
            },
            "description": "Minimum suspicion_score for open clusters."
          }
        ],
        "responses": {
          "200": {
            "description": "Active high-suspicion STS pairs inside anchorages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StsActivitiesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized \u2014 missing or invalid Bearer token."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mantis/illegal-anchoring": {
      "get": {
        "tags": [
          "mantis"
        ],
        "summary": "Get heuristic illegal-anchoring candidates.",
        "description": "v3 heuristic using stopped/stale **Class-A large vessels** only (AIS `shipType` 70\u201389). KEEP vessels inside the **restricted-limit** polygon and/or **parent / watch polygons** from `polygons.py`. EXCLUDE vessels inside any **Excl** carve-out (a hole inside a parent polygon, e.g. Singapore South Anchorage Excl1). Parent polygons such as Singapore East / Western OPL / South are used as watch zones. OFAC labels (`sanctionsMatch`) are attached; keep/drop rules are unchanged. Not a legal determination.\n\n**This Swagger page:** `vessels` is capped at 20 samples.",
        "operationId": "getIllegalAnchoring",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Illegal-anchoring candidate vessels.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IllegalAnchoringResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized \u2014 missing or invalid Bearer token."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mantis/identity-conflict": {
      "get": {
        "tags": [
          "mantis"
        ],
        "summary": "Get re-flag / dual-MMSI identity conflicts.",
        "description": "Groups of Class-A cargo/tanker MMSIs (`shipType` 70\u201389) that resolve to **one hull** (re-flag, dual identity). Match rule: IMO plus name/callsign/dimensions, or name plus callsign/dimensions. Placeholder IMOs such as 1234567 are ignored. `detectedAt` is the **latest AIS timestamp** in the group (`ais_position.ts`, else `ais_static.ts`) \u2014 not the API wall clock. OFAC labels (`sanctionsMatch`) are attached per identity and rolled up on the group; unmatched groups are kept. Listed groups are sorted first.\n\nOptional `maxDistanceM` keeps only groups whose last positions are still that close.\n\n**This Swagger page:** `groups` and `identities` are each capped at 20 samples.",
        "operationId": "getIdentityConflicts",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "maxDistanceM",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "description": "If set, keep only groups whose last-known positions are closer than this (metres). Groups without two positions are dropped."
          }
        ],
        "responses": {
          "200": {
            "description": "Identity-conflict groups (same hull, multiple MMSIs).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityConflictResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized \u2014 missing or invalid Bearer token."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/mantis/vessel-timeline": {
      "get": {
        "tags": [
          "mantis"
        ],
        "summary": "Get derived activity timeline for one vessel.",
        "description": "Returns discrete activity/events for a vessel between two datetimes from PostgreSQL derived tables: zone visits (`ais_vesselinzone`), restricted zone visits (`ais_vesselinrestrictzone`), stop/low-speed activities (`ais_vesselmovementactivities`), slow-move activities (`ais_vesselslowmoveactivities`), and static identity changes (`ais_static_evt`). Does not include raw AIS position points \u2014 use **GET /mantis/vessel-track** for full track replay.\n\n**This Swagger page:** `events` is capped at 20 samples.",
        "operationId": "getVesselTimeline",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "mmsi",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Maritime Mobile Service Identity of the vessel.",
            "example": 533000123
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Range start (ISO 8601, UTC recommended).",
            "example": "2026-06-10T00:00:00Z"
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "Range end (ISO 8601, UTC recommended).",
            "example": "2026-06-12T00:00:00Z"
          }
        ],
        "responses": {
          "200": {
            "description": "Ordered list of derived vessel activity events.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VesselTimelineResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing/invalid query parameters or date range.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized \u2014 missing or invalid Bearer token."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
...
  }
}